+++ banzai/main.py
@@ -13,49 +13,18 @@ import os
 import sys
 import traceback
 import logging
-import operator
+import copy
 
 from kombu import Exchange, Connection, Queue
 from kombu.mixins import ConsumerMixin
 
 from banzai import settings
+from banzai import dbs, preview, logs
 from banzai.context import PipelineContext
-import banzai.images
-from banzai import bias, dark, flats, trim, photometry, astrometry, qc, logs
-from banzai import dbs
-from banzai import crosstalk, gain, mosaic, bpm
-from banzai import preview
-from banzai.qc import pointing
 from banzai.utils import image_utils, date_utils
-from banzai.context import TelescopeCriterion
-from banzai import logs
-import copy
+import banzai.images
 
 logger = logging.getLogger(__name__)
-
-ORDERED_STAGES = [qc.HeaderSanity,
-                  qc.ThousandsTest,
-                  qc.SaturationTest,
-                  bias.OverscanSubtractor,
-                  crosstalk.CrosstalkCorrector,
-                  gain.GainNormalizer,
-                  mosaic.MosaicCreator,
-                  bpm.BPMUpdater,
-                  trim.Trimmer,
-                  bias.BiasSubtractor,
-                  dark.DarkSubtractor,
-                  flats.FlatDivider,
-                  qc.PatternNoiseDetector,
-                  photometry.SourceDetector,
-                  astrometry.WCSSolver,
-                  pointing.PointingTest]
-
-IMAGING_CRITERIA = [TelescopeCriterion('camera_type', operator.contains, 'FLOYDS', exclude=True),
-                    TelescopeCriterion('camera_type', operator.contains, 'NRES', exclude=True)]
-
-SCHEDULABLE_CRITERIA = [TelescopeCriterion('schedulable', operator.eq, True)]
-
-PREVIEW_ELIGIBLE_SUFFIXES = ['e00.fits', 's00.fits', 'b00.fits', 'd00.fits', 'f00.fits']
 
 RAW_PATH_CONSOLE_ARGUMENT = {'args': ["--raw-path"],
                              'kwargs': {'dest': 'raw_path', 'default': '/archive/engineering',
@@ -86,9 +55,9 @@ def get_stages_todo(last_stage=None, extra_stages=None):
     if last_stage is None:
         last_index = None
     else:
-        last_index = ORDERED_STAGES.index(last_stage) + 1
+        last_index = settings.ORDERED_STAGES.index(last_stage) + 1
 
-    stages_todo = ORDERED_STAGES[:last_index] + extra_stages
+    stages_todo = settings.ORDERED_STAGES[:last_index] + extra_stages
     return stages_todo
 
 
@@ -100,7 +69,6 @@ def parse_args(selection_criteria, extra_console_arguments=None,
 
     parser.add_argument("--processed-path", default='/archive/engineering',
                         help='Top level directory where the processed data will be stored')
-
     parser.add_argument("--log-level", default='debug', choices=['debug', 'info', 'warning',
                                                                  'critical', 'fatal', 'error'])
     parser.add_argument('--post-to-archive', dest='post_to_archive', action='store_true',
@@ -136,7 +104,7 @@ def parse_args(selection_criteria, extra_console_arguments=None,
     logs.set_log_level(args.log_level)
 
     if not args.ignore_schedulability:
-        selection_criteria += SCHEDULABLE_CRITERIA
+        selection_criteria += settings.SCHEDULABLE_CRITERIA
 
     pipeline_context = PipelineContext(args, selection_criteria, **kwargs)
 
@@ -205,54 +173,50 @@ def parse_directory_args(pipeline_context, raw_path, selection_criteria, extra_c
 
 
 def make_master_bias(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path, ['BIAS'], last_stage=trim.Trimmer,
-                      extra_stages=[bias.BiasMasterLevelSubtractor, bias.BiasComparer, bias.BiasMaker],
-                      log_message='Making Master BIAS', calibration_maker=True)
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.BIAS_IMAGE_TYPES, last_stage=settings.BIAS_LAST_STAGE,
+                      extra_stages=settings.BIAS_EXTRA_STAGES, log_message='Making Master Bias', calibration_maker=True)
 
 
 def make_master_dark(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path, ['DARK'], last_stage=bias.BiasSubtractor,
-                      extra_stages=[dark.DarkNormalizer, dark.DarkComparer, dark.DarkMaker],
-                      log_message='Making Master Dark', calibration_maker=True)
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.DARK_IMAGE_TYPES, last_stage=settings.DARK_LAST_STAGE,
+                      extra_stages=settings.DARK_EXTRA_STAGES, log_message='Making Master Dark', calibration_maker=True)
 
 
 def make_master_flat(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path, ['SKYFLAT'], last_stage=dark.DarkSubtractor,
-                      log_message='Making Master Flat', calibration_maker=True,
-                      extra_stages=[flats.FlatNormalizer, qc.PatternNoiseDetector, flats.FlatComparer, flats.FlatMaker])
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.FLAT_IMAGE_TYPES, last_stage=settings.FLAT_LAST_STAGE,
+                      extra_stages=settings.FLAT_EXTRA_STAGES, log_message='Making Master Flat', calibration_maker=True)
 
 
 def reduce_science_frames(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path, ['EXPOSE', 'STANDARD'])
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.SCIENCE_IMAGE_TYPES)
 
 
 def reduce_single_science_frame():
     extra_console_arguments = [{'args': ['--filename'],
                                 'kwargs': {'dest': 'filename', 'help': 'Name of file to process'}}]
-    pipeline_context, raw_path = parse_directory_args(None, None, IMAGING_CRITERIA,
+    pipeline_context, raw_path = parse_directory_args(None, None, settings.IMAGING_CRITERIA,
                                                       extra_console_arguments=extra_console_arguments)
     process_single_frame(pipeline_context, raw_path, pipeline_context.filename)
 
 
 def reduce_experimental_frames(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path, ['EXPERIMENTAL'])
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.EXPERIMENTAL_IMAGE_TYPES)
 
 
 def reduce_trailed_frames(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path, ['TRAILED'])
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.TRAILED_IMAGE_TYPES)
 
 
 def preprocess_sinistro_frames(pipeline_context=None, raw_path=None):
-    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, IMAGING_CRITERIA)
-    process_directory(pipeline_context, raw_path,
-                      image_types=['EXPOSE', 'STANDARD', 'BIAS', 'DARK', 'SKYFLAT', 'TRAILED', 'EXPERIMENTAL'],
-                      last_stage=mosaic.MosaicCreator)
+    pipeline_context, raw_path = parse_directory_args(pipeline_context, raw_path, settings.IMAGING_CRITERIA)
+    process_directory(pipeline_context, raw_path, settings.SINISTRO_IMAGE_TYPES,
+                      last_stage=settings.SINISTRO_LAST_STAGE)
 
 
 def reduce_night():
@@ -263,7 +227,7 @@ def reduce_night():
                                 'kwargs': {'dest': 'rawpath_root', 'default': '/archive/engineering',
                                            'help': 'Top level directory with raw data.'}}]
 
-    pipeline_context = parse_args(IMAGING_CRITERIA, extra_console_arguments=extra_console_arguments,
+    pipeline_context = parse_args(settings.IMAGING_CRITERIA, extra_console_arguments=extra_console_arguments,
                                   parser_description='Reduce all the data from a site at the end of a night.')
 
     # Ping the configdb to get telescopes
@@ -273,53 +237,58 @@ def reduce_night():
         logger.error('Could not connect to the configdb.')
         logger.error(e)
 
-    timezone = dbs.get_timezone(pipeline_context.site, db_address=pipeline_context.db_address)
+    try:
+        timezone = dbs.get_timezone(pipeline_context.site, db_address=pipeline_context.db_address)
+    except dbs.SiteMissingException:
+        logger.error("Site {site} not found in database {db}, exiting.".format(site=pipeline_context.site,
+                                                                               db=pipeline_context.db_address),
+                     extra_tags={'site': pipeline_context.site})
+        return
 
     telescopes = dbs.get_telescopes_at_site(pipeline_context.site,
                                             db_address=pipeline_context.db_address,
                                             ignore_schedulability=pipeline_context.ignore_schedulability)
 
-    if timezone is not None:
-        # If no dayobs is given, calculate it.
-        if pipeline_context.dayobs is None:
-            dayobs = date_utils.get_dayobs(timezone=timezone)
-        else:
-            dayobs = pipeline_context.dayobs
+    # If no dayobs is given, calculate it.
+    if pipeline_context.dayobs is None:
+        dayobs = date_utils.get_dayobs(timezone=timezone)
+    else:
+        dayobs = pipeline_context.dayobs
 
-        # For each telescope at the given site
-        for telescope in telescopes:
-            raw_path = os.path.join(pipeline_context.rawpath_root, pipeline_context.site,
-                                    telescope.instrument, dayobs, 'raw')
+    # For each telescope at the given site
+    for telescope in telescopes:
+        raw_path = os.path.join(pipeline_context.rawpath_root, pipeline_context.site,
+                                telescope.instrument, dayobs, 'raw')
 
-            # Run the reductions on the given dayobs
-            try:
-                make_master_bias(pipeline_context=pipeline_context, raw_path=raw_path)
-            except Exception as e:
-                logger.error(e)
-            try:
-                make_master_dark(pipeline_context=pipeline_context, raw_path=raw_path)
-            except Exception as e:
-                logger.error(e)
-            try:
-                make_master_flat(pipeline_context=pipeline_context, raw_path=raw_path)
-            except Exception as e:
-                logger.error(e)
-            try:
-                reduce_science_frames(pipeline_context=pipeline_context, raw_path=raw_path)
-            except Exception as e:
-                logger.error(e)
+    # Run the reductions on the given dayobs
+    try:
+        make_master_bias(pipeline_context=pipeline_context, raw_path=raw_path)
+    except Exception as e:
+        logger.error(e)
+    try:
+        make_master_dark(pipeline_context=pipeline_context, raw_path=raw_path)
+    except Exception as e:
+        logger.error(e)
+    try:
+        make_master_flat(pipeline_context=pipeline_context, raw_path=raw_path)
+    except Exception as e:
+        logger.error(e)
+    try:
+        reduce_science_frames(pipeline_context=pipeline_context, raw_path=raw_path)
+    except Exception as e:
+        logger.error(e)
 
 
 def get_preview_stages_todo(image_suffix):
-    if image_suffix == 'b00.fits':
-        stages = get_stages_todo(last_stage=trim.Trimmer,
-                                 extra_stages=[bias.BiasMasterLevelSubtractor, bias.BiasComparer])
-    elif image_suffix == 'd00.fits':
-        stages = get_stages_todo(last_stage=bias.BiasSubtractor,
-                                 extra_stages=[dark.DarkNormalizer, dark.DarkComparer])
-    elif image_suffix == 'f00.fits':
-        stages = get_stages_todo(last_stage=dark.DarkSubtractor,
-                                 extra_stages=[flats.FlatNormalizer, qc.PatternNoiseDetector, flats.FlatComparer])
+    if image_suffix in settings.BIAS_SUFFIXES:
+        stages = get_stages_todo(last_stage=settings.BIAS_LAST_STAGE,
+                                 extra_stages=settings.BIAS_EXTRA_STAGES_PREVIEW)
+    elif image_suffix in settings.DARK_SUFFIXES:
+        stages = get_stages_todo(last_stage=settings.DARK_LAST_STAGE,
+                                 extra_stages=settings.DARK_EXTRA_STAGES_PREVIEW)
+    elif image_suffix in settings.FLAT_SUFFIXES:
+        stages = get_stages_todo(last_stage=settings.FLAT_LAST_STAGE,
+                                 extra_stages=settings.FLAT_EXTRA_STAGES_PREVIEW)
     else:
         stages = get_stages_todo()
     return stages
@@ -335,7 +304,9 @@ def run_preview_pipeline():
                                {'args': ['--queue-name'],
                                 'kwargs': {'dest': 'queue_name', 'default': 'preview_pipeline',
                                            'help': 'Name of the queue to listen to from the fits exchange.'}}]
-    pipeline_context = parse_args(IMAGING_CRITERIA, parser_description='Reduce LCO imaging data in real time.',
+
+    pipeline_context = parse_args(settings.IMAGING_CRITERIA,
+                                  parser_description='Reduce LCO imaging data in real time.',
                                   extra_console_arguments=extra_console_arguments, preview_mode=True)
 
     # Need to keep the amqp logger level at least as high as INFO,
@@ -395,7 +366,7 @@ class PreviewModeListener(ConsumerMixin):
         message.ack()  # acknowledge to the sender we got this message (it can be popped)
 
         is_eligible_for_preview = False
-        for suffix in PREVIEW_ELIGIBLE_SUFFIXES:
+        for suffix in settings.PREVIEW_ELIGIBLE_SUFFIXES:
             if suffix in path:
                 is_eligible_for_preview = True
                 image_suffix = suffix